+++ src/containers/NewContractMember/NewContractMemberSelfConsumptionChooser.jsx
@@ -3,8 +3,8 @@ import { useTranslation } from 'react-i18next'
 import Grid from '@mui/material/Grid'
 
 import Chooser from '../../components/NewChooser'
-import { HelperText } from '../../components/InputField'
 import Typography from '@mui/material/Typography'
+import InputTitle from '../../components/InputTitle'
 
 import SolarPowerOutlinedIcon from '@mui/icons-material/SolarPowerOutlined'
 
@@ -18,19 +18,19 @@ const NewContractMemberSelfConsumptionChooser = (props) => {
   const handleSelfconsumptionQuestion = (value) => {
     setFieldValue('has_selfconsumption', value)
   }
-    
+
   const options = [
     {
       id: 'selfconsumption-on',
       icon: <SolarPowerOutlinedIcon sx={iconRequirements} />,
-      textHeader: t('GURB_SELFCONSUMPTION_YES_HEADER'),
-      textBody: t('GURB_SELFCONSUMPTION_YES_BODY')
+      textHeader: t('SELFCONSUMPTION_YES_HEADER'),
+      textBody: t('SELFCONSUMPTION_YES_BODY')
     },
     {
       id: 'selfconsumption-off',
       icon: <SolarPowerOutlinedIcon sx={iconOffRequirements} />,
-      textHeader: t('GURB_SELFCONSUMPTION_NO_HEADER'),
-      textBody: t('GURB_SELFCONSUMPTION_NO_BODY')
+      textHeader: t('SELFCONSUMPTION_NO_HEADER'),
+      textBody: t('SELFCONSUMPTION_NO_BODY')
     }
   ]
 
@@ -38,17 +38,30 @@ const NewContractMemberSelfConsumptionChooser = (props) => {
     <Grid container spacing={2}>
       <Grid item xs={12}>
         <Grid item xs={12}>
-            <Typography variant="headline3">{t('SELFCONSUMPTION_TITLE')}</Typography>
+          <Typography variant="headline3">
+            {t('SELFCONSUMPTION_TITLE')}
+          </Typography>
         </Grid>
         <Grid item xs={12}>
-            <Typography variant="body.sm.regular" color="secondary.dark">{t('RECOMMENDATION_SUBTITLE')}</Typography>
+          <Typography variant="body.md.regular" color="secondary.dark">
+            {t('RECOMMENDATION_SUBTITLE')}
+          </Typography>
+        </Grid>
+      </Grid>
+
+      <Grid item xs={12}>
+        <Grid item xs={12} sx={{ mt: 2 }}>
+          <InputTitle
+            variant="subtitle4"
+            text={t('SELFCONSUMPTION_ACTIVE')}
+            required={true}
+          />
         </Grid>
       </Grid>
       <Grid item xs={12}>
-          <Typography variant="subtitle4">{t('POTENCIA_A_CONTRACTAR')}</Typography>        
-      </Grid>
-      <Grid item xs={12}>
-          <Typography variant="subtitle2">{t('SELFCONSUMPTION_HELPER')}</Typography>        
+        <Typography variant="body.md.regular" color="secondary.dark">
+          {t('SELFCONSUMPTION_HELPER')}
+        </Typography>
       </Grid>
       <Grid item xs={12}>
         <Chooser
@@ -58,15 +71,13 @@ const NewContractMemberSelfConsumptionChooser = (props) => {
           handleChange={handleSelfconsumptionQuestion}
         />
       </Grid>
-      <Grid item xs={12}>
-        <HelperText
-          helperText={t('GURB_SELFCONSUMPTION_HELPER')}
-          iconHelper={true}
-        />
-      </Grid>
-      <Grid item xs={12}>
-          <Typography variant="subtitle2">{t('SELFCONSUMPTION_WARNING')}</Typography>        
-      </Grid>
+      {values.has_selfconsumption === 'selfconsumption-on' && (
+        <Grid item xs={12}>
+          <Typography variant="body.md.regular" color="secondary.dark">
+            {t('SELFCONSUMPTION_WARNING')}
+          </Typography>
+        </Grid>
+      )}
     </Grid>
   )
 }