+++ src/containers/NewContractMember/NewContractMemberSummary.jsx
@@ -3,7 +3,10 @@ import { useState, useEffect } from 'react'
 import { useParams } from 'react-router-dom'
 import ReviewTable from '../../components/review/ReviewTable'
 import ReviewPricesTable from '../../components/review/ReviewPrices'
-import InputTitle from '../../components/InputTitle'
+
+import EditOutlinedIcon from '@mui/icons-material/EditOutlined'
+import { buttonGurbLight } from '../../containers/Gurb/gurbTheme'
+import Button from '@mui/material/Button'
 
 import Typography from '@mui/material/Typography'
 import Grid from '@mui/material/Grid'
@@ -14,7 +17,7 @@ import LocalPhoneOutlinedIcon from '@mui/icons-material/LocalPhoneOutlined'
 import CreditCardOutlinedIcon from '@mui/icons-material/CreditCardOutlined'
 import FormControlLabel from '@mui/material/FormControlLabel'
 import Checkbox from '@mui/material/Checkbox'
-import TermsDialog from '../Gurb/components/TermsDialog'
+import TermsDialog from '../../components/TermsDialog'
 import LegalText from '../../components/LegalText'
 
 import { iconRequirements } from '../../themes/commonStyles'
@@ -26,6 +29,7 @@ import Loading from '../../components/Loading'
 import PlaceOutlinedIcon from '@mui/icons-material/PlaceOutlined'
 import SettingsOutlinedIcon from '@mui/icons-material/SettingsOutlined'
 import DescriptionOutlinedIcon from '@mui/icons-material/DescriptionOutlined'
+import LocalOfferOutlinedIcon from '@mui/icons-material/LocalOfferOutlined'
 
 const TARIFF_INDEXED = 'indexed'
 
@@ -48,27 +52,28 @@ const NewContractMemberSummary = (props) => {
   const [loading, setLoading] = useState(false)
   const [prices, setPrices] = useState({})
   const [openGeneralTermsDialog, setOpenGeneralTermsDialog] = useState(false)
+  const [showReviewLinks, setShowReviewLinks] = useState(false)
 
   const handleChangePrivacyPolicy = (event) => {
     const checked = event.target.checked
-    setFieldValue('new_member.privacy_policy_accepted', checked)
-    setFieldTouched('new_member.privacy_policy_accepted', true)
+    setFieldValue('privacy_policy_accepted', checked)
+    setFieldTouched('privacy_policy_accepted', true)
   }
 
   const handleAcceptGeneralTerms = () => {
-    setFieldValue('new_member.generic_conditions_accepted', true)
+    setFieldValue('generic_conditions_accepted', true)
     setOpenGeneralTermsDialog(false)
   }
 
   const handleCloseGeneralTerms = () => {
-    setFieldValue('new_member.generic_conditions_accepted', false)
+    setFieldValue('generic_conditions_accepted', false)
     setOpenGeneralTermsDialog(false)
   }
 
   const handleChangeStatutes = (event) => {
     const checked = event.target.checked
-    setFieldValue('new_member.statutes_accepted', checked)
-    setFieldTouched('new_member.statutes_accepted', true)
+    setFieldValue('statutes_accepted', checked)
+    setFieldTouched('statutes_accepted', true)
   }
 
   const languages = {
@@ -80,59 +85,77 @@ const NewContractMemberSummary = (props) => {
 
   const legalReviewFields = {
     icon: <PersonIcon sx={iconRequirements} />,
-    title: t('GURB_REVIEW_HOLDER_TITLE'),
+    title: t('REVIEW_HOLDER_TITLE'),
     field: [
       {
         reviewLabel: t('BUSINESS_NAME'),
         reviewValue: values?.new_member?.name,
-        step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+        step: showReviewLinks
+          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+          : null
       },
       {
         reviewLabel: t('CIF'),
         reviewValue: values?.new_member?.nif,
-        step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['IDENTIFY_MEMBER']
+        step: showReviewLinks
+          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['IDENTIFY_MEMBER']
+          : null
       },
       {
         reviewLabel: t('PROXY'),
         reviewValue: `${values?.new_member?.proxyname} (${values?.new_member?.proxynif})`,
-        step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+        step: showReviewLinks
+          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+          : null
       },
       {
-        reviewLabel: t('GURB_REVIEW_SUPPLY_POINT_LABEL_ADDRESS'),
+        reviewLabel: t('REVIEW_SUPPLY_POINT_LABEL_ADDRESS'),
         reviewValue: `${values?.address?.street} ${values?.address?.number}`,
-        step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+        step: showReviewLinks
+          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+          : null
       },
       {
-        reviewLabel: t('GURB_REVIEW_SUPPLY_POINT_LABEL_CITY'),
+        reviewLabel: t('REVIEW_SUPPLY_POINT_LABEL_CITY'),
         reviewValue: values?.address?.city?.name,
-        step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+        step: showReviewLinks
+          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+          : null
       }
     ]
   }
 
   const physicalReviewFields = {
     icon: <PersonIcon sx={iconRequirements} />,
-    title: t('GURB_REVIEW_HOLDER_TITLE'),
+    title: t('REVIEW_HOLDER_TITLE'),
     field: [
       {
-        reviewLabel: t('GURB_REVIEW_HOLDER_LABEL_NAME'),
+        reviewLabel: t('REVIEW_HOLDER_LABEL_NAME'),
         reviewValue: `${values?.new_member?.name} ${values?.new_member?.surname1} ${values?.new_member?.surname2}`,
-        step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+        step: showReviewLinks
+          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+          : null
       },
       {
-        reviewLabel: t('GURB_REVIEW_HOLDER_LABEL_NIF'),
+        reviewLabel: t('REVIEW_HOLDER_LABEL_NIF'),
         reviewValue: values?.new_member?.nif,
-        step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['IDENTIFY_MEMBER']
+        step: showReviewLinks
+          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['IDENTIFY_MEMBER']
+          : null
       },
       {
-        reviewLabel: t('GURB_REVIEW_HOLDER_LABEL_PHONE'),
+        reviewLabel: t('REVIEW_HOLDER_LABEL_PHONE'),
         reviewValue: `(${values?.new_member?.phone_code}) ${values?.new_member?.phone}`,
-        step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+        step: showReviewLinks
+          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+          : null
       },
       {
-        reviewLabel: t('GURB_REVIEW_HOLDER_LABEL_EMAIL'),
+        reviewLabel: t('REVIEW_HOLDER_LABEL_EMAIL'),
         reviewValue: values?.new_member?.email,
-        step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+        step: showReviewLinks
+          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+          : null
       }
     ]
   }
@@ -151,8 +174,8 @@ const NewContractMemberSummary = (props) => {
           }
         ]
       },
-      values?.new_member?.is_physical == 'physic-person'
-        ? physicalReviewFields
+      values?.new_member?.person_type == 'legal-person'
+        ? legalReviewFields
         : physicalReviewFields
     ],
     [
@@ -163,22 +186,30 @@ const NewContractMemberSummary = (props) => {
           {
             reviewLabel: t('CUPS_LABEL'),
             reviewValue: values?.cups,
-            step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['SUPPLY_POINT']
+            step: showReviewLinks
+              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['SUPPLY_POINT']
+              : null
           },
           {
-            reviewLabel: t('GURB_REVIEW_SUPPLY_POINT_LABEL_ADDRESS'),
+            reviewLabel: t('REVIEW_SUPPLY_POINT_LABEL_ADDRESS'),
             reviewValue: `${values?.address?.street} ${values?.address?.number}`,
-            step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['SUPPLY_INFO']
+            step: showReviewLinks
+              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['SUPPLY_INFO']
+              : null
           },
           {
-            reviewLabel: t('GURB_REVIEW_SUPPLY_POINT_LABEL_CITY'),
+            reviewLabel: t('REVIEW_SUPPLY_POINT_LABEL_CITY'),
             reviewValue: values?.address?.city?.name,
-            step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['SUPPLY_INFO']
+            step: showReviewLinks
+              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['SUPPLY_INFO']
+              : null
           },
           {
             reviewLabel: 'CNAE',
-            reviewValue: values?.supply_point.CNAE,
-            step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['SUPPLY_INFO']
+            reviewValue: values?.supply_point.cnae,
+            step: showReviewLinks
+              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['SUPPLY_INFO']
+              : null
           }
         ]
       },
@@ -193,14 +224,16 @@ const NewContractMemberSummary = (props) => {
           {
             reviewLabel: t('FARE'),
             reviewValue: values?.contract.tariff_mode,
-            step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+            step: showReviewLinks
+              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+              : null
           },
           {
             reviewLabel: t('POWER'),
             reviewValue: t('CURRENT')
           },
           {
-            reviewValue: t('GURB_REVIEW_TECHNICAL_DETAILS_FOOTER')
+            reviewValue: t('REVIEW_TECHNICAL_DETAILS_FOOTER')
           }
         ]
       }
@@ -208,46 +241,55 @@ const NewContractMemberSummary = (props) => {
     [
       {
         icon: <LocalPhoneOutlinedIcon sx={iconRequirements} />,
-        title: t('GURB_REVIEW_CONTACT_INFORMATION_TITLE'),
+        title: t('REVIEW_CONTACT_INFORMATION_TITLE'),
         field: [
           {
-            reviewLabel: t('GURB_REVIEW_HOLDER_LABEL_PHONE'),
+            reviewLabel: t('REVIEW_HOLDER_LABEL_PHONE'),
             reviewValue: `(${values?.new_member?.phone_code}) ${values?.new_member?.phone}`,
-            step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+            step: showReviewLinks
+              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+              : null
           },
           {
-            reviewLabel: t('GURB_REVIEW_HOLDER_LABEL_EMAIL'),
+            reviewLabel: t('REVIEW_HOLDER_LABEL_EMAIL'),
             reviewValue: values?.new_member?.email,
-            step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+            step: showReviewLinks
+              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+              : null
           },
           {
             reviewLabel: t('LANGUAGE'),
             reviewValue: languages[values?.new_member?.language],
-            step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+            step: showReviewLinks
+              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+              : null
           }
         ]
       },
       {
         icon: <CreditCardOutlinedIcon sx={iconRequirements} />,
-        title: t('GURB_REVIEW_PAYMENT_DATA_TITLE'),
+        title: t('REVIEW_PAYMENT_DATA_TITLE'),
         field: [
           {
             reviewLabel: t('REVIEW_PAYMENT_DATA'),
             reviewValue: t('REVIEW_PAYMENT_DATA_QUANTITY'),
-            step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['PAYMENT_INFO']
+            step: showReviewLinks
+              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['PAYMENT_INFO']
+              : null
           },
           {
-            reviewLabel: t('GURB_REVIEW_PAYMENT_DATA_LABEL_IBAN'),
+            reviewLabel: t('REVIEW_PAYMENT_DATA_LABEL_IBAN'),
             reviewValue: values?.new_member?.iban,
-            step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['PAYMENT_INFO']
+            step: showReviewLinks
+              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['PAYMENT_INFO']
+              : null
           },
           {
             reviewLabel: t('VOLUNTARY_CENT'),
-            reviewValue:
-              values?.voluntary_donation == 'voluntary-donation-on'
-                ? t('YES')
-                : t('NO'),
-            step: NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['DONATION']
+            reviewValue: values?.voluntary_donation ? t('YES') : t('NO'),
+            step: showReviewLinks
+              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['DONATION']
+              : null
           }
         ]
       }
@@ -256,7 +298,7 @@ const NewContractMemberSummary = (props) => {
 
   const reviewPrices = [
     {
-      title: t('GURB_REVIEW_PRICES_ENERGY_TITLE'),
+      title: t('REVIEW_PRICES_ENERGY_TITLE'),
       field: 'energia'
     },
     {
@@ -264,7 +306,7 @@ const NewContractMemberSummary = (props) => {
       field: 'gkwh'
     },
     {
-      title: t('GURB_REVIEW_PRICES_POWER_TITLE'),
+      title: t('REVIEW_PRICES_POWER_TITLE'),
       field: 'potencia'
     },
     { title: t('AUTOCONSUM'), field: 'energia_consumida' }
@@ -304,7 +346,13 @@ const NewContractMemberSummary = (props) => {
         setLoading(false)
         console.error(error)
       })
-  }, [values])
+  }, [
+    values.contract.power,
+    values.contract.power_type,
+    values.new_member.nif,
+    values.supply_point.cnae,
+    values?.supply_point_address?.city?.id
+  ])
 
   const handleCheckboxChange = async (event, fieldName) => {
     let value = event.target.checked
@@ -316,36 +364,84 @@ const NewContractMemberSummary = (props) => {
     <Loading />
   ) : (
     <Grid container spacing={4}>
-      <Grid item xs={12}>
-        <Typography variant="headline3">{t('MEMBER_PAGE_SUMMARY')}</Typography>
-      </Grid>
-      <Grid item xs={12}>
-        <InputTitle text={t('MEMBER_PAGE_SUMMARY_QUESTION')} />
+      <Grid
+        item
+        xs={12}
+        sx={{
+          mt: 4,
+          display: 'flex',
+          justifyContent: 'space-between',
+          alignItems: 'center'
+        }}>
+        <Typography variant="headline4">
+          {t('MEMBER_PAGE_SUMMARY_QUESTION')}
+        </Typography>
+        {!showReviewLinks && (
+          <Button
+            id="edit_button"
+            size="small"
+            sx={{
+              ...buttonGurbLight,
+              minWidth: 'auto',
+              width: 'auto',
+              // padding: '6px 15px',
+              fontSize: '0.90rem',
+              textTransform: 'none'
+              // height: 36
+            }}
+            startIcon={<EditOutlinedIcon fontSize="medium" />}
+            onClick={() => setShowReviewLinks(true)}>
+            {t('EDIT_DATA')}
+          </Button>
+        )}
       </Grid>
       <Grid item xs={12}>
         <ReviewTable tableFields={reviewFields} />
       </Grid>
-      {isTariffIndexed && (
-        <>
-          <Grid item xs={12}>
-            <ReviewPricesTable reviewPrices={reviewPrices} prices={prices} />
+      {!isTariffIndexed ? (
+        <Grid item xs={12}>
+          <ReviewPricesTable reviewPrices={reviewPrices} prices={prices} />
+        </Grid>
+      ) : (
+        <Grid container spacing={0}>
+          <Grid item xs={2} sm={1}>
+            <LocalOfferOutlinedIcon sx={iconRequirements} />
           </Grid>
-          <Grid item xs={12}>
-            <Divider sx={{ my: 2 }} />
+          <Grid item xs={10} sm={11}>
+            <Grid container spacing={3}>
+              <Grid item xs={12}>
+                <Typography variant="body.md.bold">
+                  {t('REVIEW_PRICES_POWER_TITLE')}
+                </Typography>
+              </Grid>
+              <Grid item xs={12}>
+                <Typography
+                  variant="body.sm"
+                  dangerouslySetInnerHTML={{
+                    __html: t('REVIEW_PRICES_NOTICE_INDEXED', {
+                      url: t('REVIEW_PRICES_NOTICE_INDEXED_URL')
+                    })
+                  }}></Typography>
+              </Grid>
+            </Grid>
           </Grid>
-        </>
+        </Grid>
       )}
 
       <Grid item xs={12}>
+        <Divider sx={{ my: 2 }} />
+      </Grid>
+
+      <Grid item xs={12}>
         <Typography variant="body2">
-          {t('GURB_SUMMARY_OTHER_CONCEPTS_TITLE')}
+          {t('SUMMARY_OTHER_CONCEPTS_TITLE')}
         </Typography>
       </Grid>
       <Grid item xs={12}>
         <Typography
           variant="body2"
           dangerouslySetInnerHTML={{
-            __html: t('GURB_SUMMARY_OTHER_CONCEPTS_BODY')
+            __html: t('SUMMARY_OTHER_CONCEPTS_BODY')
           }}
         />
       </Grid>
@@ -365,7 +461,7 @@ const NewContractMemberSummary = (props) => {
           control={
             <Checkbox
               data-cy="privacy_policy"
-              checked={values?.new_member.privacy_policy_accepted || false}
+              checked={values?.privacy_policy_accepted || false}
               onChange={handleChangePrivacyPolicy}
             />
           }
@@ -385,13 +481,14 @@ const NewContractMemberSummary = (props) => {
         <FormControlLabel
           control={
             <Checkbox
-              checked={values?.new_member?.generic_conditions_accepted || false}
+              data-cy="generic_conditions_accepted"
+              checked={values?.generic_conditions_accepted || false}
               onChange={(event) => {
                 const checked = event.target.checked
                 if (checked) {
                   setOpenGeneralTermsDialog(true)
                 } else {
-                  setFieldValue('new_member.generic_conditions_accepted', false)
+                  setFieldValue('generic_conditions_accepted', false)
                 }
               }}
             />
@@ -409,7 +506,7 @@ const NewContractMemberSummary = (props) => {
       </Grid>
       <Grid item xs={12}>
         <TermsDialog
-          title={t('GENERAL_TERMS_TITLE')}
+          title={t('CONTRACTUAL_PACKAGE')}
           open={openGeneralTermsDialog}
           onAccept={handleAcceptGeneralTerms}
           onClose={handleCloseGeneralTerms}
@@ -429,7 +526,7 @@ const NewContractMemberSummary = (props) => {
           control={
             <Checkbox
               data-cy="statutes"
-              checked={values?.new_member.statutes_accepted || false}
+              checked={values?.statutes_accepted || false}
               onChange={handleChangeStatutes}
             />
           }
@@ -457,12 +554,9 @@ const NewContractMemberSummary = (props) => {
             control={
               <Checkbox
                 data-cy="comercial_info_accepted"
-                checked={values?.new_member?.comercial_info_accepted}
+                checked={values?.comercial_info_accepted}
                 onChange={(event) => {
-                  handleCheckboxChange(
-                    event,
-                    'new_member.comercial_info_accepted'
-                  )
+                  handleCheckboxChange(event, 'comercial_info_accepted')
                 }}
               />
             }