+++ src/containers/NewMember/PhysicalMemberPersonalData.jsx
@@ -22,27 +22,6 @@ const languages = {
   gl_ES: 'Galego'
 }
 
-const gender_options = {
-  woman: t('GENDER_WOMAN'),
-  man: t('GENDER_MAN'),
-  genderfluid: t('GENDER_FLUID'),
-  nonbinary: t('GENDER_NON_BINARY'),
-  others: t('GENDER_OTHERS'),
-  nottosay: t('GENDER_NOT_SAY')
-}
-
-const how_meet_us_options = {
-  O1: t('HOW_MEET_US_OPTION_1'),
-  O2: t('HOW_MEET_US_OPTION_2'),
-  O3: t('HOW_MEET_US_OPTION_3'),
-  O4: t('HOW_MEET_US_OPTION_4'),
-  O5: t('HOW_MEET_US_OPTION_5'),
-  O6: t('HOW_MEET_US_OPTION_6'),
-  O7: t('HOW_MEET_US_OPTION_7'),
-  O8: t('HOW_MEET_US_OPTION_8'),
-  O9: t('HOW_MEET_US_OPTION_9')
-}
-
 const PhysicalMemberPersonalData = (props) => {
   const {
     values,
@@ -54,6 +33,27 @@ const PhysicalMemberPersonalData = (props) => {
     setErrors
   } = props
   const { i18n, t } = useTranslation()
+
+  const gender_options = {
+    woman: t('GENDER_WOMAN'),
+    man: t('GENDER_MAN'),
+    genderfluid: t('GENDER_FLUID'),
+    nonbinary: t('GENDER_NON_BINARY'),
+    others: t('GENDER_OTHERS'),
+    nottosay: t('GENDER_NOT_SAY')
+  }
+
+  const how_meet_us_options = {
+    O1: t('HOW_MEET_US_OPTION_1'),
+    O2: t('HOW_MEET_US_OPTION_2'),
+    O3: t('HOW_MEET_US_OPTION_3'),
+    O4: t('HOW_MEET_US_OPTION_4'),
+    O5: t('HOW_MEET_US_OPTION_5'),
+    O6: t('HOW_MEET_US_OPTION_6'),
+    O7: t('HOW_MEET_US_OPTION_7'),
+    O8: t('HOW_MEET_US_OPTION_8'),
+    O9: t('HOW_MEET_US_OPTION_9')
+  }
 
   const handleChange = useHandleChange(setFieldValue)
   const handleBlur = useHandleBlur(setFieldTouched)